Reorganization of correlation handler.
