+++ handler/correlation/correlation_test.go
@@ -8,34 +8,6 @@ import (
 	"testing"
 )
 
-func TestGetOrCreateID(t *testing.T) {
-	tests := []struct {
-		in      string
-		wantNew bool
-	}{
-		{
-			in:      "",
-			wantNew: true,
-		},
-		{
-			in:      "nonsense",
-			wantNew: false,
-		},
-	}
-
-	for k, v := range tests {
-		got := getOrCreateID(v.in)
-
-		if v.wantNew == true && got == v.in {
-			t.Errorf("%v: wanted new UUID but got old one", k)
-		}
-
-		if !v.wantNew == true && got != v.in {
-			t.Errorf("%v: wanted old UUID but got new one", k)
-		}
-	}
-}
-
 func TestCorrelationNewID(t *testing.T) {
 	var gotCorrelationHeaderInside bool
